Add code injection for copart
Turns out just adding the vin to the Copart URL doesn't work, and the sitedoesn't actually submit the text in the search field. We need to do a little code injection to put the text in the field, tell the page it's been updated, then click submit. I think this might be doable in two lines, if the script that is called on compositionend is putting returned validated text into a hidden field, but I'm lazy.